[fix] 修正拼字問題
