--- PersonalTools/src/main/java/com/robbie/personaltools/font/api/cheatmeal/deletecheatmealitem/DeleteCheatMealItemPresentation.java
@@ -1,15 +1,0 @@
-package com.robbie.personaltools.font.api.cheatmeal.deletecheatmealitem;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service("deleteCheatMealItemPresentation")
-@RequiredArgsConstructor
-public class DeleteCheatMealItemPresentation {
-
-  private final DeleteCheatMealItemFlow deleteCheatMealItemFlow;
-
-  public void execute(Long cheatMealId) throws Exception {
-    this.deleteCheatMealItemFlow.execute(cheatMealId);
-  }
-}
