+++ PersonalTools/src/main/java/com/robbie/personaltools/front/api/cheatmeal/updatecheatmeal/model/UpdateCheatMealItemRequest.java
@@ -1,0 +1,20 @@
+package com.robbie.personaltools.front.api.cheatmeal.updatecheatmeal.model;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class UpdateCheatMealItemRequest {
+  /** 作弊餐點名稱 */
+  private String name;
+
+  /** 作弊餐點等級 */
+  private Integer level;
+
+  /** 作弊餐點類別 */
+  private String category;
+
+  /** 作弊餐點消耗額度 */
+  private Integer point;
+}
